backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
qdrant-client==1.6.9
meilisearch==0.28.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
starlette-context==0.3.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
python-magic==0.4.27
tika==2.6.0
requests==2.31.0
pillow==10.1.0
opencv-python==4.8.1.78
pypdf2==3.0.1
celery==5.3.4
flower==1.2.0